requirements-dev.txt
-r requirements.txt

# add packages needed only for development

django-debug-toolbar==2.2
black==19.10b0
flake8==3.8.3
coverage==5.2